ChatBot/requirements.txt
streamlit>=1.32.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.36.2
torch>=2.0.0
